+++ packages/abap-deploy-config-writer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/abap-deploy-config-writer
 
+## 0.0.59
+
+### Patch Changes
+
+-   Updated dependencies [3734fe8]
+    -   @sap-ux/btp-utils@0.16.0
+    -   @sap-ux/system-access@0.5.17
+
 ## 0.0.58
 
 ### Patch Changes