+++ packages/ui5-library-sub-generator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/ui5-library-sub-generator
 
+## 0.0.21
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.7.6
+-   @sap-ux/nodejs-utils@0.1.1
+
 ## 0.0.20
 
 ### Patch Changes